# contracts/v1/intelligence.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "InternalServiceKey": {
        "in": "header",
        "name": "X-Internal-Service-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Spring이 인증한 입력의 모델 추정·RAG·랭킹·데이터 분석",
    "title": "Property Concierge 내부 AI·데이터 서비스",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/data/market/districts": {
      "get": {
        "description": "기존 클라이언트를 위한 호환 API. 신규 화면은 regions/summary를 사용한다.",
        "operationId": "district_market_summary_internal_v1_data_market_districts_get",
        "parameters": [
          {
            "in": "query",
            "name": "sido",
            "required": false,
            "schema": {
              "default": "서울특별시",
              "title": "Sido",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 60,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "property_type",
            "required": false,
            "schema": {
              "default": "all",
              "pattern": "^(all|apartment|row_house|detached|officetel|non_residential|industrial|land)$",
              "title": "Property Type",
              "type": "string"
            }
          },
          {
            "description": "만원 단위",
            "in": "query",
            "name": "budget_max",
            "required": false,
            "schema": {
              "default": 0,
              "description": "만원 단위",
              "minimum": 0,
              "title": "Budget Max",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-internal-service-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "District Market Summary",
        "tags": [
          "data-analysis"
        ]
      }
    },
    "/internal/v1/data/market/regions": {
      "get": {
        "operationId": "list_regions_internal_v1_data_market_regions_get",
        "parameters": [
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "default": "sido",
              "pattern": "^(sido|sigungu|eupmyeondong|eup_myeon_dong|ri)$",
              "title": "Level",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parent_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 10,
                  "minLength": 10,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Code"
            }
          },
          {
            "in": "header",
            "name": "x-internal-service-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Regions",
        "tags": [
          "data-analysis"
        ]
      }
    },
    "/internal/v1/data/market/regions/summary": {
      "get": {
        "operationId": "region_market_summary_internal_v1_data_market_regions_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "region_code",
            "required": true,
            "schema": {
              "maxLength": 10,
              "minLength": 10,
              "title": "Region Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_level",
            "required": false,
            "schema": {
              "default": "sigungu",
              "pattern": "^(sigungu|eup_myeon_dong)$",
              "title": "Group Level",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 60,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "property_type",
            "required": false,
            "schema": {
              "default": "all",
              "pattern": "^(all|apartment|row_house|detached|officetel|non_residential|industrial|land)$",
              "title": "Property Type",
              "type": "string"
            }
          },
          {
            "description": "만원 단위",
            "in": "query",
            "name": "budget_max",
            "required": false,
            "schema": {
              "default": 0,
              "description": "만원 단위",
              "minimum": 0,
              "title": "Budget Max",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "area_min_sqm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 100000,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Min Sqm"
            }
          },
          {
            "in": "query",
            "name": "area_max_sqm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 100000,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Max Sqm"
            }
          },
          {
            "in": "query",
            "name": "min_build_year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 1800,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Build Year"
            }
          },
          {
            "in": "query",
            "name": "max_build_year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 1800,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Build Year"
            }
          },
          {
            "in": "header",
            "name": "x-internal-service-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Service-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Region Market Summary",
        "tags": [
          "data-analysis"
        ]
      }
    },
...
  },
  "security": [
    {
      "InternalServiceKey": []
    }
  ]
}